chore: remove legacy claw app surfaces
