+++ packages/browseros-agent/apps/claw-app/components/sidebar/SidebarNavigation.tsx
@@ -1,10 +1,4 @@
-import {
-  Bot,
-  LayoutDashboard,
-  PlugZap,
-  ScrollText,
-  ShieldCheck,
-} from 'lucide-react'
+import { LayoutDashboard, PlugZap, ScrollText } from 'lucide-react'
 import type { ComponentType, SVGProps } from 'react'
 import { NavLink, useLocation } from 'react-router'
 import {
@@ -24,35 +18,12 @@ interface NavItem {
   icon: ComponentType<SVGProps<SVGSVGElement>>
 }
 
-const legacyUi = import.meta.env.VITE_COCKPIT_LEGACY_UI === '1'
-
-/**
- * Agents and Governance route entries are only registered in App.tsx
- * when `VITE_COCKPIT_LEGACY_UI=1`, so the sidebar drops them from the
- * v2 default to avoid dead links. Same flag, same source of truth.
- */
-const v2NavItems: NavItem[] = [
+const navItems: NavItem[] = [
   { name: 'Cockpit', to: '/', icon: LayoutDashboard },
   { name: 'MCP', to: '/mcp', icon: PlugZap },
   { name: 'Audit', to: '/audit', icon: ScrollText },
 ]
 
-const legacyNavItems: NavItem[] = [
-  { name: 'Cockpit', to: '/', icon: LayoutDashboard },
-  { name: 'Agents', to: '/agents', icon: Bot },
-  { name: 'Governance', to: '/governance', icon: ShieldCheck },
-  { name: 'MCP', to: '/mcp', icon: PlugZap },
-  { name: 'Audit', to: '/audit', icon: ScrollText },
-]
-
-const navItems: NavItem[] = legacyUi ? legacyNavItems : v2NavItems
-
-/**
- * `/agents/new` lives under `/agents` so the Agents item should stay
- * highlighted when the new-agent wizard is open. `/` is otherwise an
- * exact match. The rest use simple `startsWith` so future nested
- * routes (e.g. `/governance/audit`) still highlight correctly.
- */
 function isNavItemActive(item: NavItem, pathname: string): boolean {
   if (item.to === '/') return pathname === '/'
   return pathname.startsWith(item.to)
