+++ packages/browseros-agent/apps/claw-app/modules/api/activity.hooks.ts
@@ -4,7 +4,7 @@ import type { RunStatus } from '@/lib/status'
 export interface ActivityRow {
   id: string
   agentLabel: string
-  /** Color the agent dot. Matches the per-agent color on AgentRow. */
+  /** Color for the per-agent dot in recent activity. */
   color: string
   status: Extract<
     RunStatus,
@@ -16,8 +16,8 @@ export interface ActivityRow {
   site?: string
   when: string
   /**
-   * Run id used to route a done row's Replay link to
-   * `/governance/audit/:runId/replay`. Only required on done rows.
+   * Run id used to route a done row's Replay link. Only required on
+   * done rows.
    */
   runId?: string
   /** Total tool dispatches recorded against this tab. Surfaces as a badge. */
