--- packages/browseros-agent/apps/claw-app/screens/live-run/live-run.data.ts
@@ -1,9 +1,0 @@
-import { useNavigate, useParams } from 'react-router'
-import { useRun } from '@/modules/api/run.hooks'
-
-export function useLiveRunData() {
-  const { runId = '' } = useParams<{ runId: string }>()
-  const { data: run, isLoading } = useRun({ variables: { runId } })
-  const navigate = useNavigate()
-  return { run, runId, isLoading, navigate }
-}
